requirements.txt
discord.py
python-dotenv
Quart
requests
